service/src/test/java/org/nmdp/service/epitope/gl/filter/ArsAlleleFilterTest.java
/*

    epitope-service  T-cell epitope group matching service for DPB1 locus.
    Copyright (c) 2014-2015 National Marrow Donor Program (NMDP)
    
    This library is free software; you can redistribute it and/or modify it
    under the terms of the GNU Lesser General Public License as published
    by the Free Software Foundation; either version 3 of the License, or (at
    your option) any later version.
    
    This library is distributed in the hope that it will be useful, but WITHOUT
    ANY WARRANTY; with out even the implied warranty of MERCHANTABILITY or
    FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public
    License for more details.
    
    You should have received a copy of the GNU Lesser General Public License
    along with this library;  if not, write to the Free Software Foundation,
    Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA.
    
    > http://www.gnu.org/licenses/lgpl.html

*/

package org.nmdp.service.epitope.gl.filter;

import static org.hamcrest.Matchers.equalTo;
import static org.junit.Assert.assertThat;

import org.junit.Test;
import org.nmdp.service.epitope.gl.filter.ArsAlleleFilter;

public class ArsAlleleFilterTest {

	ArsAlleleFilter filter = new ArsAlleleFilter();
	
	@Test
	public void testApply() throws Exception {
		assertThat(filter.apply("DPB1*01:01:01/DPB1*01:01:02"), equalTo("DPB1*01:01/DPB1*01:01"));
	}

}
